Refactoring of FitModel and some other changes
FitModel parameters are now an instance attribute. This allows two modes
of using `FitModel`.

This first is by overriding `_func` this allows you to have a better typed
experience. The FitParameter metadata is attached as annotations to the
arguments of `_func` and automatically extracted from there.

The second mode is to for dynamic parameter sets. The parameters are
passed explicitly to the `FitModel` constructor and `func` is overridden.

The renaming utility is re-cast as a FitModel with dynamic parameters and
to me at least feels less magic now than it did before because it can do
all the work as regular methods called a wrapped `FitModel` object rather
than having to do dynamic class creation.

In addition to updating the tests to instantiate `FitModel` objects there
are a couple of tweaks to the imports to make them work with my tooling.