Attempt to fix data with multi-dimensional x-axis data.
I tried to make a new class that inherits from Model but supports data with two x-axis degrees of freedom.

There is a lot I like about this approach, but making it work with Fitter is messy.

Fitter makes a lot of assumptions (e.g. how we sort and rescale data) which work
simply with single-dimensional data but become messy with multi-dimensional data.

It wasn't obvious to me how to make this work in the more general case without
(a) making the single-dimensional use-case (which is, after all, the primary
purpose of this toolkit!) less conveient  and (b) introducing a lot of additional
complexity.

Maybe some day I'll come back to this approach, but for now I'm going to go back
to fitting the data hierarchically...