Make Model2D inherit from Model
Removes the parameter name remapping functionality from Model2D in favour of
letting the user wrap the model in a WrappedModel. This avoids duplicating
functionality and lets each container focs on doing "one thing well".

Using an additional container to do the parameter renaming does include
a bit of additional overhead (more dictionary look ups) but looking at
timings on the tests, it's not significant (running the test a few times
any increases in runtime are lost within the noise).