Avoid using inner functions for scaling (#201)
We need to pickle the model to send it to the ARTIQ applet. However, if
the model includes anonymous/non-module-level functions, then pickling
will serialise the function's code, rather than a reference to it.

This is a big problem when the client and server are on different Python
versions, as now we can't load the function on the client! Instead we
move the functions to the top-level, and wrap them with partial(...)
instead.